@@ -17,6 +17,7 @@ import {
   isNumber,
 } from '../../../../Common/utils/jsUtils';
 import { getUnderlyingType } from '../../../../../shared/utils/graphqlSchemaUtils';
+import { TableDefinition } from '../../../../Common/utils/v1QueryUtils';
 
 export type ArgValueKind = 'column' | 'static';
 export type ArgValue = {
@@ -264,7 +265,17 @@ const getTypedArgValueInput = (argValue: ArgValue, type: string) => {
   };
 };
 
-export const getRemoteRelPayload = (relationship: RemoteRelationship) => {
+export type RemoteRelationshipPayload = {
+  name: string;
+  remote_schema: string;
+  remote_field: Record<string, RemoteRelationshipFieldServer>;
+  hasura_fields: string[];
+  table: TableDefinition;
+};
+
+export const getRemoteRelPayload = (
+  relationship: RemoteRelationship
+): RemoteRelationshipPayload => {
   const hasuraFields: string[] = [];
   const getRemoteFieldArguments = (field: RemoteField) => {
     const getArgumentObject = (depth: number, parent?: string) => {
@@ -325,7 +336,7 @@ export const getRemoteRelPayload = (relationship: RemoteRelationship) => {
   return {
     name: relationship.name,
     remote_schema: relationship.remoteSchema,
-    remote_field: getRemoteFieldObject(0),
+    remote_field: getRemoteFieldObject(0) || {},
     hasura_fields: hasuraFields
       .map(f => f.substr(1))
       .filter((v, i, s) => s.indexOf(v) === i),
