@@ -3,11 +3,13 @@ import { Route, IndexRoute, IndexRedirect } from 'react-router';
 
 import { connect } from 'react-redux';
 
-import { App, Main, PageNotFound } from 'components';
-
 import globals from './Globals';
 
+import { App, Main, PageNotFound } from 'components';
+
 import validateLogin from './utils/validateLogin';
+
+import { requireAsyncGlobals } from './components/App/Actions';
 
 import { composeOnEnterHooks } from 'utils/router';
 
@@ -15,11 +17,11 @@ import { loadMigrationStatus } from './components/Main/Actions';
 
 import { dataRouterUtils } from './components/Services/Data';
 
-import { eventRouterUtils } from './components/Services/EventTrigger';
-
 import { getRemoteSchemaRouter } from './components/Services/RemoteSchema';
 
 import { getActionsRouter } from './components/Services/Actions';
+
+import { getEventsRouter } from './components/Services/Events';
 
 import generatedApiExplorer from './components/Services/ApiExplorer/ApiExplorer';
 
@@ -74,13 +76,6 @@ const routes = store => {
   const requireSchema = _dataRouterUtils.requireSchema;
   const dataRouter = _dataRouterUtils.makeDataRouter;
 
-  const _eventRouterUtils = eventRouterUtils(
-    connect,
-    store,
-    composeOnEnterHooks
-  );
-  const eventRouter = _eventRouterUtils.makeEventRouter;
-
   const remoteSchemaRouter = getRemoteSchemaRouter(
     connect,
     store,
@@ -88,6 +83,8 @@ const routes = store => {
   );
 
   const actionsRouter = getActionsRouter(connect, store, composeOnEnterHooks);
+
+  const eventsRouter = getEventsRouter(connect, store, composeOnEnterHooks);
 
   const uiKitRouter = globals.isProduction ? null : (
     <Route
@@ -103,7 +100,14 @@ const routes = store => {
   );
 
   return (
-    <Route path="/" component={App} onEnter={validateLogin(store)}>
+    <Route
+      path="/"
+      component={App}
+      onEnter={composeOnEnterHooks([
+        validateLogin(store),
+        requireAsyncGlobals(store),
+      ])}
+    >
       <Route path="login" component={generatedLoginConnector(connect)} />
       <Route
         path=""
@@ -138,9 +142,9 @@ const routes = store => {
             <Route path="about" component={aboutConnector(connect)} />
           </Route>
           {dataRouter}
-          {eventRouter}
           {remoteSchemaRouter}
           {actionsRouter}
+          {eventsRouter}
           {uiKitRouter}
         </Route>
       </Route>
